refactor(domains-services): refactor useDeploymentStatus
